requirements.txt
streamlit
pillow
torch
timm
matplotlib
numpy
scikit-learn
opencv-python
seaborn
torchvision
opencv-python
captum